Add MVP version of sg, the Sourcegraph developer command
See ./dev/sg/README.md and the #dev-experience Slack channel for details.

This adds a basic MVP version of `sg` that can be used to boot up the development environment.

Why merge this now? Because I want to be able to open small PRs that contribute to it vs. iterating on the same PR. And I also want to nerd-snipe others into working on it.