sg: add ability to define `checks` per `commandset` (#23020)
This allows us to check whether services are running before we boot up
the environment. This should reduce the number of "why is service X failing?".
At least that's the hope.

The changes to the config file format are backwards compatible. See the
`UnmarshalYAML` method on `Commandset`.

It also updates the checks to make sure that they work for users of the
docker-compose setup.